resolvers: opath: emulate MS_NOSYMFOLLOW
For openat2, the kernel will handle MS_NOSYMFOLLOW for us, but for the
O_PATH resolver we need to emulate it (just like fs.protected_symlinks).

The slightly complicated bit is that rustix doesn't fully expose all the
bits we need for MS_NOSYMFOLLOW (MountFlags::NOSYMFOLLOW is only exposed
for the non-libc backend, while ST_NOSYMFOLLOW is completely
unexported). This requires us to hardcode some values, at least for now.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>