*: add minimal {From,Into}<OwnedFd> tests
Our test coverage shows we never actually called these bindings in our
code -- we should probably do that.

We can't do these round-trip tests in the generic integration tests
because we implement RootImpl and HandleImpl for the owned and ref
versions of the structures (but {From,Into}<OwnedFd> can only be
implemented for the owned versions).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>