procfs: add ProcfsBase::ProcPid
This is a fairly common thing people want to do, and it doesn't cost us
much to implement it. The main benefit is that C users don't need to do
strncat(3) every time they just want to do a simple one-shot operation.

Part of this change requires us to modify the enum constants for
pathrs_proc_base_t, which is a breaking change but nobody is using
libpathrs yet. To future-proof this as much as possible, we reserve the
top 2^32 values of the space (real PIDs are not that large and currently
cannot be larger than 2^22, and this seems unlikely to change any time
soon -- especially since glibc uses int16_t for pid_t).

Unfortunately, it turns out that CGo has a known issue when dealing with
u64 enum values (among other things) so it is necessary to add some
hacks in the Go bindings to make sure they compile after this change.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>