python bindings: add ProcfsBase::ProcPid support
This is a bit more straightforward than the Go support -- we just needed
to add a single new function to create the new ProcfsBase values.

However, we also need to update the filters we were using for
<pathrs.h>. Previously we silently stripped all macros (as cffi has very
limited support for macros), but now that __PATHRS_PROC_TYPE_* are plain
constant macros we need update them to be more selective about macro
stripping.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>