python bindings: switch pyproject to SPDX license string
PEP 639 requires that we switch to SPDX license strings (and PyPI will
stop accepting packages using the old system in 2026).

Unfortunately, this requires a minimum of setuptools 77.0.3[1], which
means we can no longer support Python 3.8 -- but this is a fairly old
release from 2019. Older distributions that want to package libpathrs
can easily patch around this if necessary.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>