utils: sysctl: genericise sysctl_read_line using FromStr
FromStr is implemented for String, so we can just use the same
str::parse function with String to get back the same string value. In
addition to nicely unifying the logic, This feels a bit more ergonomic
as well.

The only downside is that FromStr::Err is Infallible in that impl (for
obvious reasons), so we need to add a dummy entry to ErrorImpl to make
the types happy. This variant can never actually be constructed, for the
same reason Infallible cannot be constructed.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>