syscalls: fix -EBADF handling for rustix as_raw_fd check
rustix 1.0 added support for -EBADF as a RawFd value (rustix::fs::ABS),
but HotfixRustixFd would still return an error if you tried to use it
(meaning that BADFD didn't actually work for our syscalls).

Also, rustix has only allow-listed AT_FDCWD and -EBADF -- all other
dirfd values will still cause panics. So this "hotfix" is going to
basically be permanent, and so we should rename this to something that
other than "hotfix".

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>