utils: fd: add fdinfo parser
/proc/self/fdinfo includes very useful information -- most notably the
mnt_id for any file handle. This will allow us to fetch the mnt_id of
open files even on kernels without STATX_MNT_ID.

The alternative approach would be to parse /proc/self/mountinfo, but
/proc/self/fdinfo is harder for an attacker to fake because we also
check that the "ino" field matches the actual inode number we get from
fstat(2).

Thwarting this would require the attacker to be able to correctly and
consistently guess the inode number of the file being operated on, as
well as the fd being targeted, as well as convincingly creating fake
fdinfo files using fairly restrictive gadgets like /proc/self/environ.

A sufficiently smart attacker with good knowledge of the procfs access
patterns of the victim process might be able to do it, but the attack
also needs to be very active because most inode numbers are not very
stable in procfs. /proc/self/mountinfo is much easier to attack (if you
just set it to a static file without overmounts, the overmounts are
invisible).

In addition, /proc/self/fdinfo works with detached mount fds (a-la
fsopen(2) -- which is relevant for AT_RECURISVE trees due to locked
mounts), and is _much_ easier to parse than /proc/self/mountinfo.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>