syscalls: add internal feature to force-disable openat2
This will allow us to test our code in a way that emulates kernels
without openat2 (or openat2 disabled with seccomp).

Rather than #[cfg(...)]-gating every bit of openat2 code, it is much
less messy to just make the underlying syscall wrapper return -ENOSYS
(which will then cause *OPENAT2_IS_SUPPORTED to be false). This will
also let us test -ENOSYS fallbacks.

This feature _technically_ breaks the general rule that features should
be additive[1], but the alternative would be to make this an "openat2"
feature which would need to be added to default-features (which then
would cause issues if someone used libpathrs with default features
disabled).

A key thing to notice is that this feature is implemented by only
enabling an existing if statement (a forceful -ENOSYS return), which
should assuage most of the concerns around non-additive features. Also,
this feature is hidden from the crates.io docs and is explicitly not a
supported feature, so we wouldn't support a user building libpathrs with
this feature added (why would you even want to?)

[1]: https://doc.rust-lang.org/cargo/reference/features.html#feature-unification

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>