utils: use fdinfo as final fallback for fetch_mnt_id
Now that we have a hardened /proc/thread-self/fdinfo handler, we can use
this as a final fallback when looking for the mount id of a path. Every
"struct file" in the system has an associated "mnt_id" in fdinfo, which
is a huge benefit.

This is not an ideal solution, but it should be quite difficult for an
attacker to bypass, and it critically allows us to provide complete
security for post-openat2-but-pre-STATX_MNT_ID kernels.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>