*: switch to impl Trait for most APIs
It turns out that taking "impl Trait" arguments is actually more ergonomic
in a lot of cases:

 * It makes it easier to write APIs where the caller specifying a
   generic type with ::<> (like str::parse::<T>) much easier. If every
   type parameter is named then you need to do ::<_, _, Foo, _>() which
   is awful.

 * It avoids the need to define a generic type for each argument even if
   they implement the same trait (<P1: AsRef<Path>, P2: AsRef<Path>).
   This is needed to avoid annoying reference issues (and it turns out
   that we missed fixing this in a few places).

 * The rustdoc documentation generated looks much nicer -- it is far
   clearer to users what types we accept (previously, the use of generic
   type parameter names would require users to look at the where clause
   to see what Trait impl is needed).

Note that these types are still monomorphised during compilation (unlike
the very-similar-looking "&dyn Trait" arguments). It's a shame that this
appears to be fairly underutilised in Rust libraries (so much so that I
didn't realise this was a feature -- I thought "impl Trait" always
referred to a single concrete type without needing to spell out its
name).

I also removed a few ill-advised AsRef<str> arguments.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>