utils: add RawProcfsRoot helper
Rather than passing around Option<BorrowedFd<'_>> between ProcfsHandle
helper functions, we should have a separate type to better indicate that
this is a bit of a compromise solution.

This should also make it harder to accidentally pass the wrong file
descriptor as the wrong argument when we extend RawProcfsRoot usage.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>