docs: sort kernel support table by kernel version
We probably should've always done this, but openat2(2) was originally
put first since it's probably the most important feature and was the
first thing I added.

But to users I suspect that ordering by minimum kernel version is far
easier to follow and provides more useful information ("given a kernel
version, above this line is supported, below this line is not").

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>