tests: use cargo-hack to test feature powerset
This lets us test libpathrs with openat2 support completely disabled (in
principle as though we were running on an older kernel, though in
practice the openat2 code is entirely #[cfg(...)]'d out).

In order to reduce the (exponential) explosition of feature
combinations, we disable separately testing purely-additive features
(like capi) for the integration tests. However, the linters will run for
all feature combinations.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>