syscalls: add internal feature to force-disable statx
Now that we have the /proc/thread-self/fdinfo fallback, we need to test
it by disabling STATX_MNT_ID. Unfortunately, some of our core procfs
tests overmount /proc/self/fdinfo, so we need to update those.

Note that (unlike openat2), it really isn't necessary to re-run the
entire test suite when we disable statx -- all we really care about is
the procfs stuff.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>