gha: add commit status for binary size
Some of our users will care a fair bit about our binary size (in
particular when static linking), so it's a good idea to keep abreast of
this when reviewing PRs regularly.

Annoyingly, GITHUB_TOKEN does not have the necessary permissions to set
commit statuses when running in a pull request, and there is no way to
change this (the "permissions" configuration only affects actions run
from the main repo). The official solution for this is to create your
own User Access Tokens, but I cannot create such tokens for the openSUSE
organisation -- we might need to move libpathrs to my personal account.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>